+++ src/bufferutils.js
@@ -42,9 +42,9 @@ function reverseBuffer(buffer) {
 }
 exports.reverseBuffer = reverseBuffer;
 function cloneBuffer(buffer) {
-  const clone = Buffer.alloc(buffer.length);
+  const clone = Buffer.allocUnsafe(buffer.length);
   buffer.copy(clone);
-  return buffer;
+  return clone;
 }
 exports.cloneBuffer = cloneBuffer;
 /**