+++ ts_src/bufferutils.ts
@@ -49,9 +49,9 @@ export function reverseBuffer(buffer: Buffer): Buffer {
 }
 
 export function cloneBuffer(buffer: Buffer): Buffer {
-  const clone = Buffer.alloc(buffer.length);
+  const clone = Buffer.allocUnsafe(buffer.length);
   buffer.copy(clone);
-  return buffer;
+  return clone;
 }
 
 /**